requirements.txt
flask==2.3.3
google-cloud-vision==3.4.0
google-cloud-storage==2.10.0
python-dotenv==1.0.0
requests==2.31.0
Pillow>=10.0.0
gunicorn==21.2.0
flask-cors==4.0.0
google-cloud-aiplatform>=1.36.0 